+++ timer.h
@@ -30,7 +30,7 @@ extern volatile u8  time_5ms;
 
 
 // delay in task MENU
-extern void delay_menu(u8 len_5ms);
+extern void delay_menu(u16 len_5ms);
 
 
 #endif
